ADD: Inital octree implementation form models
