--- dependencies/include/application/Service.h
@@ -1,16 +1,0 @@
-#ifndef __SERVICE_H__
-#define __SERVICE_H__
-
-//  All the services inherit this class
-class Service
-{
-private:
-public:
-    bool initalized = true;
-
-    Service() {}
-    // Service* GetService() { return this; }
-    
-};
-
-#endif
