+++ dependencies/include/services/Service.h
@@ -1,0 +1,16 @@
+#ifndef __SERVICE_H__
+#define __SERVICE_H__
+
+//  All the services inherit this class
+class Service
+{
+private:
+public:
+    bool initalized = true;
+
+    Service() {}
+    // Service* GetService() { return this; }
+    
+};
+
+#endif
